Added support for --rt option and QuickPath feature
The command line interface now accepts a fourth unspecified option and
a specified option --rt for declaring the running time of HPCG.  If the
specified time is exactly 0 (zero), the QuickPath mode is enabled,
reducing the runtime of all verification and timing phases to their
minimum.  This mode can be used for established systems with approval
from the HPCG Benchmark leaders.